codeintel: Move gitserver CommitsUniqueToBranch into git package
